requirements.txt
langchain 
langchain-openai 
chromadb 
unstructured 
python-dotenv
langchain_community
pythainlp
unstructured